requirements.txt
Django==5.1.7
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
python-dotenv==1.0.1
sqlparse==0.5.3
asgiref==3.8.1
